+++ modules/webtools/src/cpp/RepositoryClone.hpp
@@ -35,5 +35,10 @@ void
 RepositoryClone(const std::wstring& url, const std::wstring& branch, std::wstring& localPath,
     std::wstring& errorMessage);
 //=============================================================================
+NLSWEBTOOLS_IMPEXP
+void
+RepositoryExport(const std::wstring& url, const std::wstring& branch, std::wstring& localPath,
+    std::wstring& errorMessage);
+//=============================================================================
 };
 //=============================================================================